+++ src/webgis4u/ol/control/OverviewMap.js
@@ -20,7 +20,7 @@ import * as extent from 'ol/extent';
 
 import { createElement } from '../../util/dom/createElement';
 import { getMapChildElementByClassName } from '../../util/ol/getMapChildElementByClassName';
-import { EPSG31287_ID } from '../proj/epsg31287';
+import { EPSG31287_ID } from '../proj/austria';
 
 import { CSS_CONTROL_PREFIX } from './common';
 import './OverviewMap/OverviewMap.scss';