+++ src/webgis4u/ol/proj.js
@@ -2,5 +2,5 @@
  * @module webgis4u/ol/proj
  */
 
-export * from './proj/epsg31287';
+export * from './proj/austria';
 export * from './proj/reproject';