+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased changes
-* Added projections for the austrian area - [`epsg31287`](./src/webgis4u/ol/proj/epsg31287.js)
+* Added projections for the austrian area - [`austria`](./src/webgis4u/ol/proj/austria.js)
 * Added [`basemap`](./src/webgis4u/ol/source/basemap.js) as a source - for more information, see [basemap.at](https://www.basemap.at/)
 * Added following controls:
   * [`PanBar`](./src/webgis4u/ol/control/PanBar.js)